BuyerService merged into CustomerService
