+++ src/main/java/com/projects/quickcart/controller/ProductController.java
@@ -7,19 +7,19 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
-import com.projects.quickcart.service.BuyerService;
+import com.projects.quickcart.service.CustomerService;
 
 import jakarta.servlet.http.HttpSession;
 
 @Controller
 public class ProductController {
 	@Autowired
-	private BuyerService bService;
+	private CustomerService customerService;
 
 	@GetMapping("/products")
 	public ModelAndView productsView(@RequestParam(required = false) String category) {
 		ModelAndView mView = new ModelAndView("products");
-		mView.addObject("producs", bService.getProducts(category));
+		mView.addObject("producs", customerService.getProducts(category));
 		return mView;
 	}
 
@@ -28,11 +28,11 @@ public class ProductController {
 		ModelAndView mView = new ModelAndView("product-info");
 		var id1 = (Long) session.getAttribute("userId");
 		if (id1 != null) {
-			boolean isWishlisted = bService.isProductWishlisted(id1, id);
+			boolean isWishlisted = customerService.isProductWishlisted(id1, id);
 			mView.addObject("wishlisted", isWishlisted);
-			mView.addObject("inCart", bService.isProductInCart(id1, id));
+			mView.addObject("inCart", customerService.isProductInCart(id1, id));
 		}
-		mView.addObject("product", bService.getProduct(id));
+		mView.addObject("product", customerService.getProduct(id));
 		return mView;
 	}
 }
